+++ packages/azure/src/azure-openai-provider.ts
@@ -1,0 +1,89 @@
+import {
+  OpenAIChatLanguageModel,
+  OpenAIChatSettings,
+} from '@ai-sdk/openai/internal';
+import { loadApiKey, loadSetting } from '@ai-sdk/provider-utils';
+
+export interface AzureOpenAIProvider {
+  (
+    deploymentId: string,
+    settings?: OpenAIChatSettings,
+  ): OpenAIChatLanguageModel;
+
+  /**
+Creates an Azure OpenAI chat model for text generation.
+   */
+  chat(
+    deploymentId: string,
+    settings?: OpenAIChatSettings,
+  ): OpenAIChatLanguageModel;
+}
+
+export interface AzureOpenAIProviderSettings {
+  /**
+Name of the Azure OpenAI resource.
+     */
+  resourceName?: string;
+
+  /**
+API key for authenticating requests.
+     */
+  apiKey?: string;
+}
+
+/**
+Create an Azure OpenAI provider instance.
+ */
+export function createAzure(
+  options: AzureOpenAIProviderSettings = {},
+): AzureOpenAIProvider {
+  const getHeaders = () => ({
+    'api-key': loadApiKey({
+      apiKey: options.apiKey,
+      environmentVariableName: 'AZURE_API_KEY',
+      description: 'Azure OpenAI',
+    }),
+  });
+
+  const getResourceName = () =>
+    loadSetting({
+      settingValue: options.resourceName,
+      settingName: 'resourceName',
+      environmentVariableName: 'AZURE_RESOURCE_NAME',
+      description: 'Azure OpenAI resource name',
+    });
+
+  const createChatModel = (
+    deploymentName: string,
+    settings: OpenAIChatSettings = {},
+  ) =>
+    new OpenAIChatLanguageModel(deploymentName, settings, {
+      provider: 'azure-openai.chat',
+      headers: getHeaders,
+      url: ({ path, modelId }) =>
+        `https://${getResourceName()}.openai.azure.com/openai/deployments/${modelId}${path}?api-version=2024-05-01-preview`,
+      compatibility: 'compatible',
+    });
+
+  const provider = function (
+    deploymentId: string,
+    settings?: OpenAIChatSettings,
+  ) {
+    if (new.target) {
+      throw new Error(
+        'The Azure OpenAI model function cannot be called with the new keyword.',
+      );
+    }
+
+    return createChatModel(deploymentId, settings as OpenAIChatSettings);
+  };
+
+  provider.chat = createChatModel;
+
+  return provider as AzureOpenAIProvider;
+}
+
+/**
+Default Azure OpenAI provider instance.
+ */
+export const azure = createAzure({});