+++ packages/openai/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ai-sdk/openai
 
+## 0.0.27
+
+### Patch Changes
+
+- fc9552ec: fix (@ai-sdk/azure): allow for nullish delta
+
+## 0.0.26
+
+### Patch Changes
+
+- 7530f861: fix (@ai-sdk/openai): add internal dist to bundle
+
+## 0.0.25
+
+### Patch Changes
+
+- 8b1362a7: chore (@ai-sdk/openai): expose models under /internal for reuse in other providers
+
 ## 0.0.24
 
 ### Patch Changes